requirement.txt
asgiref==3.7.2
Django==5.0
django-ckeditor==6.7.1
django-jalali==6.0.1
django-js-asset==2.2.0
jdatetime==4.1.1
mysqlclient==2.2.1
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
